eliminated all extraneous files in test directory now just runtests.jl and @testsnippet. All tests pass.
