--- test/TestPrograms/SphericalHarmonics.jl
@@ -1,96 +1,0 @@
-@memoize function P(l, m, z)
-    if l == 0 && m == 0
-        return 1.0
-    elseif l == m
-        return (1 - 2m) * P(m - 1, m - 1, z)
-    elseif l == m + 1
-        return (2m + 1) * z * P(m, m, z)
-    else
-        return ((2l - 1) / (l - m) * z * P(l - 1, m, z) - (l + m - 1) / (l - m) * P(l - 2, m, z))
-    end
-end
-export P
-
-@memoize function S(m, x, y)
-    if m == 0
-        return 0
-    else
-        return x * C(m - 1, x, y) - y * S(m - 1, x, y)
-    end
-end
-export S
-
-@memoize function C(m, x, y)
-    if m == 0
-        return 1
-    else
-        return x * S(m - 1, x, y) + y * C(m - 1, x, y)
-    end
-end
-export C
-
-function factorial_approximation(x)
-    local n1 = x
-    sqrt(2 * π * n1) * (n1 / ℯ * sqrt(n1 * sinh(1 / n1) + 1 / (810 * n1^6)))^n1
-end
-export factorial_approximation
-
-function compare_factorial_approximation()
-    for n in 1:30
-        println("n $n relative error $((factorial(big(n))-factorial_approximation(n))/factorial(big(n)))")
-    end
-end
-export compare_factorial_approximation
-
-@memoize function N(l, m)
-    @assert m >= 0
-    if m == 0
-        return sqrt((2l + 1 / (4π)))
-    else
-        # return sqrt((2l+1)/2π * factorial(big(l-m))/factorial(big(l+m)))
-        #use factorial_approximation instead of factorial because the latter does not use Stirlings approximation for large n. Get error for n > 2 unless using BigInt but if use BigInt get lots of rational numbers in symbolic result.
-        return sqrt((2l + 1) / 2π * factorial_approximation(l - m) / factorial_approximation(l + m))
-    end
-end
-export N
-
-"""l is the order of the spherical harmonic"""
-@memoize function Y(l, m, x, y, z)
-    @assert l >= 0
-    @assert abs(m) <= l
-    if m < 0
-        return N(l, abs(m)) * P(l, abs(m), z) * S(abs(m), x, y)
-    else
-        return N(l, m) * P(l, m, z) * C(m, x, y)
-    end
-end
-export Y
-
-SHFunctions(max_l, x::FD.Node, y::FD.Node, z::FD.Node) = SHFunctions(Vector{FD.Node}(undef, 0), max_l, x, y, z)
-
-function SHFunctions(shfunc, max_l, x, y, z)
-    for l in 0:max_l-1
-        for m in -l:l
-            push!(shfunc, Y(l, m, x, y, z))
-        end
-    end
-
-    return shfunc
-end
-export SHFunctions
-
-function spherical_harmonics(::FastSymbolic, model_size::Integer, x, y, z)
-    graph = FD.DerivativeGraph(SHFunctions(model_size, x, y, z))
-    return graph
-end
-
-function spherical_harmonics(package::FastSymbolic, model_size::Integer)
-    @variables x y z
-    return spherical_harmonics(package, model_size, x, y, z)
-end
-
-spherical_harmonics(model_size::Integer) = spherical_harmonics(FastSymbolic(), model_size)
-spherical_harmonics(model_size::Integer, x, y, z) = spherical_harmonics(FastSymbolic(), model_size, x, y, z)
-export spherical_harmonics
-
-
