--- test/TestPrograms/TestCode.jl
@@ -1,2 +1,0 @@
-include("Types.jl")
-include("SphericalHarmonics.jl")
